slightly improve height calculation
Width and height must be integers, and if any of them is a float, the
default value is used silently (without warning). Using the 1024px
width accidentally results the height to be an integer, but when I
changed to some other value, it did not work, and I did not understand
why.
-
Ticket: SUITEDEV-28293